added a coroutine that takes in an int which is a max_delay and it will wait for a random between 0 and max_delay

+++ python_async_function/0-basic_async_syntax.py
@@ -11,7 +11,7 @@ import asyncio
 import random
 
 
-async def wait_random(max_delay: int = 10) -> int:
+async def wait_random(max_delay: int = 10) -> float:
     """
         This function takes in a max_delay, and int argument
         and wait for a random delay between 0 and max_delay
